==> 01-DATA_RETRIEVAL.ipynb <==
# Jupyter Notebook Purpose
- Run all code cells and you will have the most current data from each resource (through a series of HTTP requests)
    - as long as the links remain active in the future
- The data will be saved in the data format from where they came and unzipped for convenience
- Also the data will be processed in later Jupyter Notebooks
    - they are numbered in the order that they should be run (ex. 01, 02, etc.)

## Group 10 Members

- ### A. Nidhi Punja - [Email](mailto:npunja@uwaterloo.ca)
- ### B. Judith Roth - [Email](mailto:j5roth@uwaterloo.ca)
- ### C. Iman Dordizadeh Basirabad - [Email](mailto:idordiza@uwaterloo.ca)
- ### D. Daniel Adam Cebula - [Email](mailto:dacebula@uwaterloo.ca)
- ### E. Cynthia Fung - [Email](mailto:c27fung@uwaterloo.ca)
- ### F. Ben Klassen - [Email](mailto:b6klasse@uwaterloo.ca)

# Group 10 Collaborators
COLLABORATORS = ["Nidhi Punja",
                 "Judith Roth",
                 "Iman Dordizadeh Basirabad",
                 "Daniel Adam Cebula",
                 "Cynthia Fung",
                 "Ben Klassen"]

# Group 10 Members
for _ in COLLABORATORS:
    print(f"Group 10 Member: {_:->30}")

# Table of Contents

## 1. [Python Dependecies](#1.-Python-Libraries-and-Dependencies[1,2,3,4,5])
___
## 2. [Folder Creation for Data Storage](#2.-Folders-for-Data-Storage)
___
## 3. [Toronto Fire Services Basic Incident Details (2011 - 2018)](#3.-Toronto-Fire-Services-Basic-Incident-Details[6,7])
### 3a. [Data Download / Saving / Unzipping](#3a.-Data-Retrieval)
### 3b. [Data Munging / Wrangling](#3b.-Data-Manipulation)
### 3c. [Initial DataFrame](#3c.-Initial-DataFrame-of-TFS-Fire-Incidents[7])
___
## 4. [Toronto Historical Weather](#4.-Toronto-Historical-Weather-and-Climate-from-the-Government-of-Canada[8,9])
### 4a. [Data Download / Saving / Unzipping](#4a.-Data-Retrieval)
### 4b. [Data Munging / Wrangling](#4b.-Data-Manipulation)
### 4c. [Initial DataFrame](#4c.-Initial-DataFrame-of-Historical-Toronto-Weather-and-Climate)
___
## 5. [Toronto Fire Services Station Locations](#5.-Toronto-Fire-Services-(TFS)-Fire-Station-Locations[6,11])
### 5a. [Data Download / Saving / Unzipping](#5a.-Data-Retrieval)
___
## 6. [References](#6.-Jupyter-Notebook-References)
___

# 1. Python Libraries and Dependencies<sup>[1,2,3,4,5]</sup>

# Python Modules for Miscellaneous reasons
from zipfile import ZipFile  # to read and write to zipped folders
import requests  # simple HTTP library for Python
import os        # portable way to use operating system functionalities
import io        # Tool for working with streams (Input/Ouput data)
import datetime  # python classes for manipulating dates and times
import dateutil  # powerful extensions to standard datetime Python module
import time      # used for time.sleep() to delay the HTTP requests ever so slightly
import re        # used for Python regex library
from create_folder import create_folder # create folder function that I have defined and placed in create_folder.py file

# DATA ANALYSIS / VISUALIZATION Python Dependencies
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
%matplotlib inline

# 2. Folders for Data Storage
- these folders will be used to store `RAW_ZIPPED/`, `RAW_UNZIPPED/`, `PROCESSED_ZIPPED/`, `PROCESSED_UNZIPPED/` data files
- it is assumed that you do not have them so I will create them for you if you do not have them

# Here are the major directory names that will hold the data / metadata
RAW_ZIPPED_DIRECTORY = create_folder(folder_name="RAW_ZIPPED")
RAW_UNZIPPED_DIRECTORY = create_folder(folder_name="RAW_UNZIPPED")
PROCESSED_ZIPPED_DIRECTORY = create_folder(folder_name="PROCESSED_ZIPPED")
PROCESSED_UNZIPPED_DIRECTORY = create_folder(folder_name="PROCESSED_UNZIPPED")

# 3. Toronto Fire Services Basic Incident Details<sup>[6,7]</sup>
- Open Data Toronto [link](https://open.toronto.ca/dataset/fire-services-basic-incident-details/)
- Year Range (2011 - 2018)

## 3a. Data Retrieval

# Create folders for fire_incidents
FIRE_RAW_ZIPPED_DIRECTORY = create_folder(folder_name=os.path.join(RAW_ZIPPED_DIRECTORY, "FIRE_INCIDENTS"))
FIRE_RAW_UNZIPPED_DIRECTORY = create_folder(folder_name=os.path.join(RAW_UNZIPPED_DIRECTORY, "FIRE_INCIDENTS"))
FIRE_PROCESSED_ZIPPED_DIRECTORY = create_folder(folder_name=os.path.join(PROCESSED_ZIPPED_DIRECTORY, "FIRE_INCIDENTS"))
FIRE_PROCESSED_UNZIPPED_DIRECTORY = create_folder(folder_name=os.path.join(PROCESSED_UNZIPPED_DIRECTORY, "FIRE_INCIDENTS"))

# Make a request for the fire incidents zipped folder
URL = r"https://ckan0.cf.opendata.inter.prod-toronto.ca/download_resource/64038657-6437-4a97-b6f7-b4caf135249f"

# this might take a while to download the zipped file (~50 MB in size)
with requests.get(URL, stream=True) as response:
    
    # location where the zipped folder will be saved
    ZIPPED_FOLDER = os.path.join(FIRE_RAW_ZIPPED_DIRECTORY, "2011-2018-fire-incidents.zip")
    
    # save the zipped folder
    with open(ZIPPED_FOLDER, "wb") as file:
        for chunk in response.iter_content(chunk_size=128):
            file.write(chunk)

# Open the downloaded zipped folder through ZipFile class
with ZipFile(ZIPPED_FOLDER, "r") as unzipped_files:
    # extract all files into the fire_unzipped_data folder
    unzipped_files.extractall(FIRE_RAW_UNZIPPED_DIRECTORY)

## 3b. Data Manipulation

# get the data dictionary (metadata)
for dirpath, subdirs, files in os.walk(FIRE_RAW_UNZIPPED_DIRECTORY):
    subdirs[:] = [d for d in subdirs if d not in '.ipynb_checkpoints']
    for x in files:
        if x.split(".")[1] == "xlsx":
            DATA_DICTIONARY_PATH = os.path.join(dirpath, x)

# walk through the fire_incidents directory
# to find the file all csv files
CSV_FILE_PATHS = []
for dirpath, subdirs, files in os.walk(FIRE_RAW_UNZIPPED_DIRECTORY):
    subdirs[:] = [d for d in subdirs if d not in '.ipynb_checkpoints']
    for x in files:
        if x.split(".")[1] == "csv":
            if x != "2011-2018-TORONTO_FIRE_INCIDENTS_WEATHER.csv":
                CSV_FILE_PATHS.append(os.path.join(dirpath, x))
CSV_FILE_PATHS.sort();
# CSV_FILE_PATHS, a list, holds the pathway to the csv files

# read all 8 .csv files with data into a dictionary
# 3 columns can be read in as datetime
df_dict = dict()
for x in CSV_FILE_PATHS:
    df_dict[os.path.basename(x)] = pd.read_csv(x, parse_dates=["TFS Alarm Time", "TFS Arrival Time", "Last TFS Unit Clear Time"])

# see the keys in the dictionary
df_dict.keys()

# merge all dataframes into 1
df = pd.concat([df_dict[os.path.basename(x)] for x in CSV_FILE_PATHS])
df.shape

# create DATETIME column that will allow us to model the
# data as a time series.
# This will help out later on
df["DATETIME"] = df["TFS Alarm Time"]
df["DATETIME"]

# 3c. Initial DataFrame of TFS Fire Incidents<sup>[7]</sup>

# See the Metadata for the Fire Incident Data
# Each Column is described as a Row
pd.set_option('max_colwidth', 400)
df_meta = pd.read_excel(DATA_DICTIONARY_PATH)
df_meta

# reset pandas default option
pd.reset_option('max_colwidth')

# Write the data to the appropriate folder
df.to_csv(
        os.path.join(
            FIRE_PROCESSED_UNZIPPED_DIRECTORY, "2011-2018_Basic_Incident_Details.csv"
                    ),
        index=False
        )

# Preview Toronto Fire Incidents DataFrame
df.sample(10)

# 4. Toronto Historical Weather and Climate from the Government of Canada<sup>[8,9]</sup>
- [link](https://climate.weather.gc.ca/)
- Documentation for links can be found [here](ftp://client_climate@ftp.tor.ec.gc.ca/Pub/)
    - more specifically [here](ftp://client_climate@ftp.tor.ec.gc.ca/Pub/Get_More_Data_Plus_de_donnees/Readme.txt)
    - go to this [folder](ftp://client_climate@ftp.tor.ec.gc.ca/Pub/Get_More_Data_Plus_de_donnees/) for more info
- I will be using Python Requests library to generate HTTP requests to obtain the data

## 4a. Data Retrieval

# Create folders for toronto_weather
WEATHER_RAW_UNZIPPED_DIRECTORY = create_folder(folder_name=os.path.join(RAW_UNZIPPED_DIRECTORY, "TORONTO_WEATHER"))
WEATHER_PROCESSED_UNZIPPED_DIRECTORY = create_folder(folder_name=os.path.join(PROCESSED_UNZIPPED_DIRECTORY, "TORONTO_WEATHER"))
WEATHER_PROCESSED_ZIPPED_DIRECTORY = create_folder(folder_name=os.path.join(PROCESSED_ZIPPED_DIRECTORY, "TORONTO_WEATHER"))
WEATHER_PROCESSED_UNZIPPED_DIRECTORY = create_folder(folder_name=os.path.join(PROCESSED_UNZIPPED_DIRECTORY, "TORONTO_WEATHER"))

### Canada Historical Weather Station Inventory<sup>[10]</sup>
- needs to be queried to determine which stations will output Historical Toronto Weather Data

# station inventory list url
URL_Stations = r"ftp://client_climate@ftp.tor.ec.gc.ca/Pub/Get_More_Data_Plus_de_donnees/Station%20Inventory%20EN.csv"

# read the csv into pandas
df_stations = pd.read_csv(URL_Stations, skiprows=3)

# get the file path to save the inventory
STATION_WEATHER_RAW_UNZIPPED_DIRECTORY = os.path.join(WEATHER_RAW_UNZIPPED_DIRECTORY, "Canada_Weather_Station_Inventory.csv")

# save the inventory
df_stations.to_csv(STATION_WEATHER_RAW_UNZIPPED_DIRECTORY, index=False)
df_stations

# longitude is -79 - -79.6 and latitude is 44-43 for GTA area
df_latitude_longitude = df_stations.loc[(df_stations["Latitude (Decimal Degrees)"] <= 44) &
                                        (df_stations["Latitude (Decimal Degrees)"] >= 43.5) &
                                        (df_stations["Longitude (Decimal Degrees)"] >= -79.6) &
                                        (df_stations["Longitude (Decimal Degrees)"] <= -79)
                                       ]
df_latitude_longitude

# select all Toronto weather stations (toronto in name)
df_toronto_stations = df_latitude_longitude.loc[(df_latitude_longitude["Name"].str.lower().str.contains("toronto", regex=False)) & 
           (df_latitude_longitude["Province"].str.lower().str.contains("ontario", regex=False))]
df_toronto_stations

# make sure the weather stations contains weather data from <2010 to >2020 dates
df_year_stations = df_toronto_stations.loc[(df_toronto_stations["First Year"] <= 2010) &
                                          (df_toronto_stations["Last Year"] >= 2020)]
df_year_stations

# lets create a list of tuples
# this will be done to generate the HTTP requests for .CSV files
years = list(range(2010, 2021))
station_name_array = df_year_stations["Name"].values
station_id_array = df_year_stations["Station ID"].values
station_years = [(station_name_array[index], x, y) for index, x in enumerate(station_id_array) for y in years]

# this is how we will index for the different station and years
station_years[0]
station_years[1:11]

station_years[11]
station_years[12:22]

station_years[22]
station_years[23:]

# dict that stores all the dataframes
df_dict = {}

for name, a, b in station_years:

    # generated URL for Toronto Historical Weather Data
    URL = (f"http://climate.weather.gc.ca/climate_data/bulk_data_e.html?"+
           f"format=csv&stationID={a}&Year={b}&Month=1&Day=14&timeframe=2&submit= Download+Data")

    with requests.Session() as s:
        # download the file
        download = s.get(URL)
        # decode it to utf-8 format
        decode = download.content.decode("utf-8")

        # file path is too long for windows so lets
        # just store it in a buffer
        with io.StringIO(decode) as f:
            # read it into a dataframe
            df = pd.read_csv(f)
            
    # a little delay not to arouse suscpicion
    time.sleep(0.02)
    
    # place dataframe in a dictionary with tuples as keys
    df_dict[(name, a, b)] = df

# lets see if all the keys are present
df_dict.keys()

# toronto city station from 2010 - 2020
df_toronto_city = df_dict[station_years[0]]
df_toronto_city = df_toronto_city.append([df_dict[x] for x in station_years[1:11]])
TORONTO_CITY_PATH = os.path.join(WEATHER_RAW_UNZIPPED_DIRECTORY, "toronto_city_2010_2020.csv")
df_toronto_city.to_csv(TORONTO_CITY_PATH, index=False)

# city_centre station from 2010 - 2020
df_city_centre = df_dict[station_years[11]]
df_city_centre = df_city_centre.append([df_dict[x] for x in station_years[12:22]])
CITY_CENTRE_PATH = os.path.join(WEATHER_RAW_UNZIPPED_DIRECTORY, "toronto_city_centre_2010_2020.csv")
df_city_centre.to_csv(CITY_CENTRE_PATH, index=False)

# north york station from 2010 - 2020
df_north_york = df_dict[station_years[22]]
df_north_york = df_north_york.append([df_dict[x] for x in station_years[23:]])
NORTH_YORK_PATH = os.path.join(WEATHER_RAW_UNZIPPED_DIRECTORY, "toronto_north_york_2010_2020.csv")
df_north_york.to_csv(NORTH_YORK_PATH, index=False)

print(df_toronto_city.shape)
print(df_city_centre.shape)
print(df_north_york.shape)

## 4b. Data Manipulation

# drop the following columns as they are mostly null for each of the 3 dataframes
df_toronto_city = df_toronto_city.drop(columns=['Longitude (x)', 'Latitude (y)',
                                'Station Name', 'Climate ID', 'Data Quality', 'Max Temp Flag',
                                'Min Temp Flag', 'Mean Temp Flag', 'Heat Deg Days Flag',
                                'Cool Deg Days Flag', 'Total Rain Flag', 'Total Snow Flag',
                                'Total Precip Flag', 'Snow on Grnd Flag', 
                                'Dir of Max Gust Flag', 'Spd of Max Gust Flag',
                                'Total Snow (cm)', 'Dir of Max Gust (10s deg)', 
                                'Spd of Max Gust (km/h)', 'Date/Time'
                                ])

df_city_centre = df_city_centre.drop(columns=['Longitude (x)', 'Latitude (y)',
                                'Station Name', 'Climate ID', 'Data Quality', 'Max Temp Flag',
                                'Min Temp Flag', 'Mean Temp Flag', 'Heat Deg Days Flag',
                                'Cool Deg Days Flag', 'Total Rain Flag', 'Total Snow Flag',
                                'Total Precip Flag', 'Snow on Grnd Flag', 
                                'Dir of Max Gust Flag', 'Spd of Max Gust Flag',
                                'Total Snow (cm)', 'Dir of Max Gust (10s deg)', 
                                'Spd of Max Gust (km/h)', 'Total Rain (mm)',
                                'Snow on Grnd (cm)', 'Date/Time'
                                ])

df_north_york = df_north_york.drop(columns=['Longitude (x)', 'Latitude (y)',
                                'Station Name', 'Climate ID', 'Data Quality', 'Max Temp Flag',
                                'Min Temp Flag', 'Mean Temp Flag', 'Heat Deg Days Flag',
                                'Cool Deg Days Flag', 'Total Rain Flag', 'Total Snow Flag',
                                'Total Precip Flag', 'Snow on Grnd Flag', 
                                'Dir of Max Gust Flag', 'Spd of Max Gust Flag',
                                'Total Snow (cm)', 'Dir of Max Gust (10s deg)', 
                                'Spd of Max Gust (km/h)', 'Date/Time'
                                ])

# for the Rain, Snow and precipitation columns set the np.nan values to 0
# makes logical sense that if there is no rain / snow you would get a null value (or 0)
values = {
    'Total Rain (mm)': 0,
    'Total Precip (mm)':0,
    'Snow on Grnd (cm)':0,
        }

# do so for all 3 dataframes
df_toronto_city = df_toronto_city.fillna(value=values)
df_city_centre = df_city_centre.fillna(value=values)
df_north_york = df_north_york.fillna(value=values)

df_toronto_city.isnull().sum()

df_city_centre.isnull().sum()

df_north_york.isnull().sum()

# functions that will append the appropriate suffix to the dataframe
def suffix_x(value, suffix="_city"):
    if (value == "Year") or (value == "Month") or (value == "Day"):
        return value
    else:
        return value + suffix
    
def suffix_y(value, suffix="_centre"):
    if (value == "Year") or (value == "Month") or (value == "Day"):
        return value
    else:
        return value + suffix
    
def suffix_z(value, suffix="_york"):
    if (value == "Year") or (value == "Month") or (value == "Day"):
        return value
    else:
        return value + suffix
    
df_toronto_city.columns = df_toronto_city.columns.map(suffix_x)
df_city_centre.columns = df_city_centre.columns.map(suffix_y)
df_north_york.columns = df_north_york.columns.map(suffix_z)

# merge all 3 dataframes together on the year, month and day
df_total = df_toronto_city.merge(df_city_centre, how="inner", on=["Year", "Month", "Day"])
df_total = df_total.merge(df_north_york, how="inner", on=["Year", "Month", "Day"])

# create a Datetime column that denotes the year, month and day
df_total["DATE"] = pd.to_datetime(dict(year=df_total["Year"], month=df_total["Month"],day=df_total["Day"]))

# get maximum and minimum and mean temperatures
df_total["MAX_TEMP"] = df_total.loc[:, ['Max Temp (°C)_city', 'Max Temp (°C)_centre', 'Max Temp (°C)_york']].max(axis=1)
df_total["MIN_TEMP"] = df_total.loc[:, ['Min Temp (°C)_city', 'Min Temp (°C)_centre', 'Min Temp (°C)_york']].min(axis=1)
df_total["MEAN_TEMP"] = df_total.loc[:, ['Mean Temp (°C)_city', 'Mean Temp (°C)_centre', 'Mean Temp (°C)_york']].mean(axis=1)

# heating degree day and cooling degree day (heating / cooling of buildings)
df_total["HDD"] = df_total.loc[:, ['Heat Deg Days (°C)_city', 'Heat Deg Days (°C)_centre', 'Heat Deg Days (°C)_york']].mean(axis=1)
df_total["CDD"] = df_total.loc[:, ['Cool Deg Days (°C)_city', 'Cool Deg Days (°C)_centre', 'Cool Deg Days (°C)_york']].mean(axis=1)

# total rain, total preciitation and snow on ground
df_total["RAIN_MM"] = df_total.loc[:, ['Total Rain (mm)_city', 'Total Rain (mm)_york']].mean(axis=1)
df_total["PRECIP_MM"] = df_total.loc[:, ['Total Precip (mm)_city', 'Total Precip (mm)_centre', 'Total Precip (mm)_york']].mean(axis=1)
df_total["SNOW_CM"] = df_total.loc[:, ['Snow on Grnd (cm)_city', 'Snow on Grnd (cm)_york']].mean(axis=1)

# drop the columns which I have aggregated
df_total = df_total.drop(columns=['Max Temp (°C)_city', 'Min Temp (°C)_city',
                               'Mean Temp (°C)_city', 'Heat Deg Days (°C)_city',
                               'Cool Deg Days (°C)_city', 'Total Rain (mm)_city',
                               'Total Precip (mm)_city', 'Snow on Grnd (cm)_city',
                               'Max Temp (°C)_centre', 'Min Temp (°C)_centre', 'Mean Temp (°C)_centre',
                               'Heat Deg Days (°C)_centre', 'Cool Deg Days (°C)_centre',
                               'Total Precip (mm)_centre', 'Max Temp (°C)_york', 'Min Temp (°C)_york',
                               'Mean Temp (°C)_york', 'Heat Deg Days (°C)_york',
                               'Cool Deg Days (°C)_york', 'Total Rain (mm)_york',
                               'Total Precip (mm)_york', 'Snow on Grnd (cm)_york',
                               'Year', 'Month', 'Day'
                                ])
df_total.head()

# some nulls are left
df_total.isnull().sum()

# lets round the numbers
df_total.loc[:, ['MAX_TEMP', 'MIN_TEMP', 'MEAN_TEMP', 'HDD', 'CDD', 'RAIN_MM', 'PRECIP_MM', 'SNOW_CM']] = (df_total
        .loc[:, ['MAX_TEMP', 'MIN_TEMP', 'MEAN_TEMP', 'HDD', 'CDD', 'RAIN_MM', 'PRECIP_MM', 'SNOW_CM']].round(2)
                                                                                                           )

# save the file to the toronto_weather folder
AGGREGATED_TORONTO_WEATHER_PATH = os.path.join(WEATHER_PROCESSED_UNZIPPED_DIRECTORY, "2010-2020_Toronto_Weather.csv")
df_total.to_csv(AGGREGATED_TORONTO_WEATHER_PATH, index=False)

## 4c. Initial DataFrame of Historical Toronto Weather and Climate

df_total.sample(10)

# 5. Toronto Fire Services (TFS) Fire Station Locations<sup>[6,11]</sup>
- Toronto Open Data [link](https://open.toronto.ca/dataset/fire-station-locations/)

# 5a. Data Retrieval

# Create folders for fire_stations
STATIONS_RAW_ZIPPED_DIRECTORY = create_folder(folder_name=os.path.join(RAW_ZIPPED_DIRECTORY, "FIRE_STATIONS"))
STATIONS_RAW_UNZIPPED_DIRECTORY = create_folder(folder_name=os.path.join(RAW_UNZIPPED_DIRECTORY, "FIRE_STATIONS"))
STATIONS_PROCESSED_ZIPPED_DIRECTORY = create_folder(folder_name=os.path.join(PROCESSED_ZIPPED_DIRECTORY, "FIRE_STATIONS"))
STATIONS_PROCESSED_UNZIPPED_DIRECTORY = create_folder(folder_name=os.path.join(PROCESSED_UNZIPPED_DIRECTORY, "FIRE_STATIONS"))

# Make a request for the fire incidents zipped folder
URL = r"https://ckan0.cf.opendata.inter.prod-toronto.ca/download_resource/daf5e0ee-cff6-4661-b308-79f95c3881e9"

# this might take a while to download the zipped file (~50 MB in size)
with requests.get(URL, stream=True) as response:
    
    # location where the zipped folder will be saved
    ZIPPED_FOLDER = os.path.join(STATIONS_RAW_ZIPPED_DIRECTORY, "2019_Toronto_Fire_Stations.zip")
    
    # save the zipped folder
    with open(ZIPPED_FOLDER, "wb") as file:
        for chunk in response.iter_content(chunk_size=128):
            file.write(chunk)

# Open the downloaded zipped folder through ZipFile class
with ZipFile(ZIPPED_FOLDER, "r") as unzipped_files:
    # extract all files into the fire_unzipped_data folder
    unzipped_files.extractall(STATIONS_RAW_UNZIPPED_DIRECTORY)

### Geopandas Python Library will need to be used at this point<sup>[12]</sup>
- convert from shapefile (.shp) to a pandas DataFrame
- requires another Conda Virtual Environment
    - this is due to conflict between other Python libraries in the Anaconda base environment

# 6. Jupyter Notebook References

[1] "Python Documentation."  *Python Software Foundation*.  [Online](https://docs.python.org/).  [Accessed August 04, 2020]

[2] G. Niemeyer.  "dateutil - powerful extensions to datetime."  *dateutil*.  [Online](https://github.com/dateutil/dateutil).  [Accessed August 04, 2020]

[3] "pandas."  *PyData*.  [Online](https://pandas.pydata.org/).  [Accessed August 04, 2020]

[4] "NumPy - The fundamental package for scientific computing with Python."  *NumPy*.  [Online](https://numpy.org/).  [Accessed August 04, 2020]

[5] "Matplotlib:  Visualization with Python."  *The Matplotlib Development team*.  [Online](https://matplotlib.org/).  [Accessed August 04, 2020]

[6] "City of Toronto Open Data Portal."  *City of Toronto*.  [Online](https://open.toronto.ca/).  [Accessed August 04, 2020]

[7] "Fire Services Basic Incident Details."  *City of Toronto*.  [Online](https://open.toronto.ca/dataset/fire-services-basic-incident-details/).  [Accessed August 04, 2020]

[8] "Historical Climate Data."  *Government of Canada*.  [Online](https://climate.weather.gc.ca/).  [Accessed August 04, 2020]

[9] "URL based procedure to automatically download data in bulk from Climate Website"  *Government of Canada*.  [Online].  *ftp://client_climate@ftp.tor.ec.gc.ca/Pub/Get_More_Data_Plus_de_donnees/Readme.txt*.  [Accessed August 04, 2020]

[10] "Station Inventory."  *Government of Canada*.  [Online].  *ftp://client_climate@ftp.tor.ec.gc.ca/Pub/Get_More_Data_Plus_de_donnees/Station%20Inventory%20EN.csv*.  [Accessed August 04, 2020]

[11] "Fire Station Locations."  *City of Toronto*.  [Online](https://open.toronto.ca/dataset/fire-station-locations/).  [Accessed August 04, 2020]

[12] "GeoPandas."  *geopandas*.  [Online](https://github.com/geopandas/geopandas).  [Accessed August 04, 2020]
