==> 06-GEOSPATIAL.ipynb <==
# Jupyter Notebook Purpose
- read the .csv.gz compressed file dataset into pandas DataFrame and perform draw total counts on interactive Google Maps
    - explore through geospatial analysis where the top 1000 counts of geospatial locations (budget concerns with API) of TFS Fire Incidents occur

## Group 10 Members

- ### A. Nidhi Punja - [Email](mailto:npunja@uwaterloo.ca)
- ### B. Judith Roth - [Email](mailto:j5roth@uwaterloo.ca)
- ### C. Iman Dordizadeh Basirabad - [Email](mailto:idordiza@uwaterloo.ca)
- ### D. Daniel Adam Cebula - [Email](mailto:dacebula@uwaterloo.ca)
- ### E. Cynthia Fung - [Email](mailto:c27fung@uwaterloo.ca)
- ### F. Ben Klassen - [Email](mailto:b6klasse@uwaterloo.ca)

# Group 10 Collaborators
COLLABORATORS = ["Nidhi Punja",
                 "Judith Roth",
                 "Iman Dordizadeh Basirabad",
                 "Daniel Adam Cebula",
                 "Cynthia Fung",
                 "Ben Klassen"]

# Group 10 Members
for _ in COLLABORATORS:
    print(f"Group 10 Member: {_:->30}")

# Table of Contents

## 1. [Python Dependecies](#1.-Python-Libraries-and-Dependencies[1,2,3,4,5,6])
### 1a. [jupyter-gmaps interactive google maps](#1a.-jupyter-gmaps-Interactive-Google-Maps-Installation-Steps)
___
## 2. [Folder Creation](#2.-Folder-Creation-for-Data-Analyses-and-Visualization)
___
## 3. [Read in the Data](#3.-Read-in-DataSet-as-a-Pandas-DataFrame)
___
## 4. [GroupBy GeoSpatial DataFrames](#4.-GroupBy-TFS-Fire-Incidents-Geospatial-Data)
___
## 5. [jupyter-gmaps Google Maps](#5.-Jupyter-Gmaps-Interactive-Google-Maps)
### 5a. [Geospatial Findings](#5a.-Interactive-Google-Maps-Findings)
___
## 6. [References](#6.-Jupyter-Notebook-References)
___

# 1. Python Libraries and Dependencies<sup>[1,2,3,4,5,6]</sup>

# Python Modules for Miscellaneous reasons
import os        # portable way to use operating system functionalities
import datetime  # python classes for manipulating dates and times
import dateutil  # powerful extensions to standard datetime Python module
import re        # used for Python regex library
from IPython.display import display # use this to see the entire DataFrame in the right format
from create_folder import create_folder # create folder function that I have defined and placed in create_folder.py file
import warnings  # suppress warnings from various Python libraries
try:
    from api_keys import google_maps_api_key # import the google maps api key (need to create Google Developer Account)
except:
    print("Google Maps API-KEY not present")

# DATA ANALYSIS / VISUALIZATION Python Dependencies
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
%matplotlib inline

## 1a. jupyter-gmaps Interactive Google Maps Installation Steps

### 1. Install jupyter-gmaps - [link](https://jupyter-gmaps.readthedocs.io/en/latest/install.html)

#### A. Use conda ```conda install -c conda-forge gmaps``` or pip ```pip install gmaps```
    - press 'y' and ENTER
#### B. ```jupyter nbextension enable --py --sys-prefix widgetsnbextension```
#### C. ```jupyter nbextension enable --py --sys-prefix gmaps```
#### D. ```jupyter labextension install @jupyter-widgets/jupyterlab-manager```
#### E. Open notebook in Jupyter Notebook / Lab again
- ```jupyter lab```
- ```jupyter notebook```

# import the jupyter-gmaps Python library
import gmaps

# configure the Google Maps API_KEY with jupyter-gmaps
gmaps.configure(api_key=google_maps_api_key)

# 2. Folder Creation for Data Analyses and Visualization
- generate a folder that will hold data analyses / visualizations

# get a connection to the major directory holding the data / metadata of the DataFrame
PROCESSED_ZIPPED_DIRECTORY = create_folder(folder_name="PROCESSED_ZIPPED")

# generate a images folder and a analyses folder to hold all relevant information
IMAGES_DIRECTORY = create_folder(folder_name="IMAGES")
ANALYSES_DIRECTORY = create_folder(folder_name="ANALYSES")

# get folders for fire incidents, toronto weather and fire station location data
FIRE_PROCESSED_ZIPPED_DIRECTORY = create_folder(folder_name=os.path.join(PROCESSED_ZIPPED_DIRECTORY, "FIRE_INCIDENTS"))
WEATHER_PROCESSED_ZIPPED_DIRECTORY = create_folder(folder_name=os.path.join(PROCESSED_ZIPPED_DIRECTORY, "TORONTO_WEATHER"))
STATIONS_PROCESSED_ZIPPED_DIRECTORY = create_folder(folder_name=os.path.join(PROCESSED_ZIPPED_DIRECTORY, "FIRE_STATIONS"))

# 3. Read in DataSet as a Pandas DataFrame
- DataSet is a compressed file (.csv.bz2)
- DataSet Metadata is a .csv file

# read in the metadata from .csv into memory
# use this metatdata to explain the columns
df_metadata = pd.read_csv(
    os.path.join(FIRE_PROCESSED_ZIPPED_DIRECTORY, "FINAL_DATASET_METADATA.csv"),
    index_col="COLUMN_NAME")

# display it
with pd.option_context('display.max_colwidth', 300):
    display(df_metadata)

# read the merged DataFrame from .csv.bz2 file into DataFrame
PATH_MERGED_CSV_BZ2 = os.path.join(FIRE_PROCESSED_ZIPPED_DIRECTORY, "FINAL_DATASET.csv.bz2")

# pandas DataFrame generated and is referenced by df variable name
df = pd.read_csv(PATH_MERGED_CSV_BZ2,
                 compression='bz2', index_col="INCIDENT_NUM", parse_dates=["DATETIME"])

# FOR GEOSPATIAL DATA THIS IS NOT ENTIRELY THE CASE
# make the columns categorical (for faster queries)
# df["CAD_TYPE"] = pd.Categorical(df["CAD_TYPE"])
# df["CAD_CALL_TYPE"] = pd.Categorical(df["CAD_CALL_TYPE"])
# df["FINAL_TYPE"] = pd.Categorical(df["FINAL_TYPE"])
# df["CALL_SOURCE"] = pd.Categorical(df["CALL_SOURCE"])
# df["NAME"] = pd.Categorical(df["NAME"])
# df["ADDRESS"] = pd.Categorical(df["ADDRESS"])
# df["WARD_NAME"] = pd.Categorical(df["WARD_NAME"])
# df["MUN_NAME"] = pd.Categorical(df["MUN_NAME"])

# display it
with pd.option_context('display.max_columns', None):
    display(df.head())

# 4. GroupBy TFS Fire Incidents Geospatial Data
- groupby by Latitude and Longitude (excluding nulls) for major / minor intersections and explore the FINAL_TYPE, CAD_CALL_TYPE columns for 2011 - 2018 years
- groupby Laittude and Longitude (excluding nulls) for TFS Fire Stations and explore FINAL_TYPE, CAD_CALL_TYPE columns for 2011 - 2018 years

# drop all rows where latitude and longitude is null
df.dropna(subset=['LATITUDE', 'LONGITUDE'], inplace=True)

# Grab the top 10 "FINAL_TYPE" categories only
top_10_FINAL_TYPE = [x for x in df["FINAL_TYPE"].value_counts().index[:10]]
df_FINAL_TYPE_ = df.loc[df["FINAL_TYPE"].isin(top_10_FINAL_TYPE), :]

# Grab the top 10 "CAD_CALL_TYPE"
top_10_CAD_CALL_TYPE = [x for x in df["CAD_CALL_TYPE"].value_counts().index[:10]]
df_CAD_CALL_TYPE_ = df.loc[df["CAD_CALL_TYPE"].isin(top_10_CAD_CALL_TYPE), :]

# no nulls should be present in "LATITUDE" and "LONGITUDE" columns
df_FINAL_TYPE_.isnull().sum()

# Create a new dataframe with only Latitude / Longitude and FINAL_TYPE
df_FINAL_TYPE = df_FINAL_TYPE_.loc[:, ["LATITUDE", "LONGITUDE", "FINAL_TYPE"]]

# apply the pandas get dummies to turn catgory column to numerical (1 or 0)
df_FINAL_TYPE_DUMMIES = pd.get_dummies(df_FINAL_TYPE)

# groupby to get a count for each supplied LATITUDE and lONGITUDE pair
df_FINAL_TYPE_GROUPBY1 = (df_FINAL_TYPE_DUMMIES.iloc[:, 0:3]
              .groupby(["LATITUDE", "LONGITUDE"])
              .count()
              .rename(columns={df_FINAL_TYPE_DUMMIES.columns[2]:"TOTAL COUNT"}))

# groupby to get a sum (get relative count of each FINAL_TYPE category)
df_FINAL_TYPE_GROUPBY2 = (df_FINAL_TYPE_DUMMIES
              .groupby(["LATITUDE", "LONGITUDE"])
              .sum())

# merge the 2 dataframes together
df_FINAL_TYPE_GROUPBY = (df_FINAL_TYPE_GROUPBY1.merge(df_FINAL_TYPE_GROUPBY2,
                                                      how="inner", left_index=True, right_index=True)
                           .sort_values(by="TOTAL COUNT", ascending=False)
                           .reset_index())

# save the dataframe as a csv file
df_FINAL_TYPE_GROUPBY.to_csv(os.path.join(ANALYSES_DIRECTORY, "Top_10_FINAL_TYPE_Geospatial.csv"),
                             index=False)

df_FINAL_TYPE_GROUPBY

# Create a new dataframe with only Latitude / Longitude of TFS Stations and FINAL_TYPE
df_FINAL_TYPE_STATION = df_FINAL_TYPE_.loc[:, ["FIRE_STATION", "LATITUDE_STATION", "LONGITUDE_STATION", "FINAL_TYPE"]]

# apply the pandas get dummies to turn catgory column to numerical (1 or 0)
df_FINAL_TYPE_STATION_DUMMIES = pd.get_dummies(df_FINAL_TYPE_STATION)

# groupby to get a count for each supplied LATITUDE and lONGITUDE pair
df_FINAL_TYPE_STATION_GROUPBY1 = (df_FINAL_TYPE_STATION_DUMMIES.iloc[:, 0:4]
              .groupby(["FIRE_STATION", "LATITUDE_STATION", "LONGITUDE_STATION"])
              .count()
              .rename(columns={df_FINAL_TYPE_STATION_DUMMIES.columns[3]:"TOTAL COUNT"}))

# groupby to get a sum (get relative count of each FINAL_TYPE category)
df_FINAL_TYPE_STATION_GROUPBY2 = (df_FINAL_TYPE_STATION_DUMMIES
              .groupby(["FIRE_STATION", "LATITUDE_STATION", "LONGITUDE_STATION"])
              .sum())

# merge the 2 dataframes together
df_FINAL_TYPE_STATION_GROUPBY = (df_FINAL_TYPE_STATION_GROUPBY1.merge(df_FINAL_TYPE_STATION_GROUPBY2,
                                                      how="inner", left_index=True, right_index=True)
                           .sort_values(by="TOTAL COUNT", ascending=False)
                           .reset_index())

# save the dataframe as a csv file
df_FINAL_TYPE_STATION_GROUPBY.to_csv(os.path.join(ANALYSES_DIRECTORY, "Top_10_FINAL_TYPE_Fire_Stations_Geospatial.csv"),
                             index=False)

df_FINAL_TYPE_STATION_GROUPBY

# Create a new dataframe with only Latitude / Longitude and CAD_CALL_TYPE
df_CAD_CALL_TYPE = df_CAD_CALL_TYPE_.loc[:, ["LATITUDE", "LONGITUDE", "CAD_CALL_TYPE"]]

# apply the pandas get dummies to turn catgory column to numerical (1 or 0)
df_CAD_CALL_TYPE_DUMMIES = pd.get_dummies(df_CAD_CALL_TYPE)

# groupby to get a count for each supplied LATITUDE and lONGITUDE pair
df_CAD_CALL_TYPE_GROUPBY1 = (df_CAD_CALL_TYPE_DUMMIES.iloc[:, 0:3]
              .groupby(["LATITUDE", "LONGITUDE"])
              .count()
              .rename(columns={df_CAD_CALL_TYPE_DUMMIES.columns[2]:"TOTAL COUNT"}))

# groupby to get a sum (get relative count of each CAD_CALL_TYPE category)
df_CAD_CALL_TYPE_GROUPBY2 = (df_CAD_CALL_TYPE_DUMMIES
              .groupby(["LATITUDE", "LONGITUDE"])
              .sum())

# merge the 2 dataframes together
df_CAD_CALL_TYPE_GROUPBY = (df_CAD_CALL_TYPE_GROUPBY1.merge(df_CAD_CALL_TYPE_GROUPBY2,
                                                      how="inner", left_index=True, right_index=True)
                           .sort_values(by="TOTAL COUNT", ascending=False)
                           .reset_index())

# save the dataframe as a csv file
df_CAD_CALL_TYPE_GROUPBY.to_csv(os.path.join(ANALYSES_DIRECTORY, "Top_10_CAD_CALL_TYPE_Geospatial.csv"),
                             index=False)

df_CAD_CALL_TYPE_GROUPBY

# Create a new dataframe with only Latitude / Longitude of TFS Stations and CAD_CALL_TYPE
df_CAD_CALL_TYPE_STATION = df_FINAL_TYPE_.loc[:, ["FIRE_STATION", "LATITUDE_STATION", "LONGITUDE_STATION", "CAD_CALL_TYPE"]]

# apply the pandas get dummies to turn catgory column to numerical (1 or 0)
df_CAD_CALL_TYPE_STATION_DUMMIES = pd.get_dummies(df_CAD_CALL_TYPE_STATION)

# groupby to get a count for each supplied LATITUDE and lONGITUDE Station pair
df_CAD_CALL_TYPE_STATION_GROUPBY1 = (df_CAD_CALL_TYPE_STATION_DUMMIES.iloc[:, 0:4]
              .groupby(["FIRE_STATION", "LATITUDE_STATION", "LONGITUDE_STATION"])
              .count()
              .rename(columns={df_CAD_CALL_TYPE_STATION_DUMMIES.columns[3]:"TOTAL COUNT"}))

# groupby to get a sum (get relative count of each CAD_CALL_TYPE category)
df_CAD_CALL_TYPE_STATION_GROUPBY2 = (df_CAD_CALL_TYPE_STATION_DUMMIES
              .groupby(["FIRE_STATION", "LATITUDE_STATION", "LONGITUDE_STATION"])
              .sum())

# merge the 2 dataframes together
df_CAD_CALL_TYPE_STATION_GROUPBY = (df_CAD_CALL_TYPE_STATION_GROUPBY1.merge(df_CAD_CALL_TYPE_STATION_GROUPBY2,
                                                      how="inner", left_index=True, right_index=True)
                           .sort_values(by="TOTAL COUNT", ascending=False)
                           .reset_index())

# save the dataframe as a csv file
df_CAD_CALL_TYPE_STATION_GROUPBY.to_csv(os.path.join(ANALYSES_DIRECTORY, "Top_10_CAD_CALL_TYPE_Fire_Stations_Geospatial.csv"),
                             index=False)

df_CAD_CALL_TYPE_STATION_GROUPBY

# 5. Jupyter Gmaps Interactive Google Maps
- from the 4 DataFrame above draw several Interactive Google Maps using Google Maps API Key

# lets only look at the top 1000 locations
df_FINAL_TYPE_TOP_1000 = df_FINAL_TYPE_GROUPBY[:1000].copy()

# get locations
locations = df_FINAL_TYPE_TOP_1000[["LATITUDE", "LONGITUDE"]]

# get total count
total_count = df_FINAL_TYPE_TOP_1000["TOTAL COUNT"]

# make gmaps figure
fig = gmaps.figure(layout={
        'width': '400px',
        'height': '400px',
        'padding': '3px',
        'border': '1px solid black'
})

# create heat_map layer
heat_layer = gmaps.heatmap_layer(locations,
                                 weights=total_count,
                                 dissipating=False,
                                 max_intensity=300,
                                 point_radius=0.005)

# add heat_layer to figure
fig.add_layer(heat_layer)

# display the figure
display(fig)

# lets only look at the top 1000 locations
df_FINAL_TYPE_STATION_GROUPBY_COPY = df_FINAL_TYPE_STATION_GROUPBY.copy()

# get locations
locations_station = df_FINAL_TYPE_STATION_GROUPBY_COPY[["LATITUDE_STATION", "LONGITUDE_STATION"]]

# get total count
total_count_station = df_FINAL_TYPE_STATION_GROUPBY_COPY["TOTAL COUNT"]

# make gmaps figure
fig_station = gmaps.figure(layout={
        'width': '400px',
        'height': '400px',
        'padding': '3px',
        'border': '1px solid black'
})

# generate popup text
info_box_template = """
<dl>
<dt>Station: {FIRE_STATION}</dt>
</dl>
<dl>
<dt>Count: {TOTAL COUNT}</dt>
</dl>
"""
count_info = [info_box_template.format(**row) for index, row in df_FINAL_TYPE_STATION_GROUPBY.iterrows()]



# create marker layer
marker_station = gmaps.symbol_layer(locations_station,
                                    fill_color="blue",
                                    stroke_color="blue",
                                    scale=2,
                                    info_box_content=count_info)

# add heat_layer and symbols to figure
fig_station.add_layer(heat_layer)
fig_station.add_layer(marker_station)

# display the figure
display(fig_station)

# close the 2 jupyter gmap figures
fig.close()
fig_station.close()

# lets only look at the top 1000 locations
df_CAD_CALL_TYPE_TOP_1000 = df_CAD_CALL_TYPE_GROUPBY[:1000].copy()

# get locations
locations = df_CAD_CALL_TYPE_TOP_1000[["LATITUDE", "LONGITUDE"]]

# get total count
total_count = df_CAD_CALL_TYPE_TOP_1000["TOTAL COUNT"]

# make gmaps figure
fig = gmaps.figure(layout={
        'width': '400px',
        'height': '400px',
        'padding': '3px',
        'border': '1px solid black'
})

# create heat_map layer
heat_layer = gmaps.heatmap_layer(locations,
                                 weights=total_count,
                                 dissipating=False,
                                 max_intensity=300,
                                 point_radius=0.005)

# add heat_layer to figure
fig.add_layer(heat_layer)

# display the figure
display(fig)

# lets only look at the top 1000 locations
df_CAD_CALL_TYPE_STATION_GROUPBY_COPY = df_CAD_CALL_TYPE_STATION_GROUPBY.copy()

# get locations
locations_station = df_CAD_CALL_TYPE_STATION_GROUPBY_COPY[["LATITUDE_STATION", "LONGITUDE_STATION"]]

# get total count
total_count_station = df_CAD_CALL_TYPE_STATION_GROUPBY_COPY["TOTAL COUNT"]

# make gmaps figure
fig_station = gmaps.figure(layout={
        'width': '400px',
        'height': '400px',
        'padding': '3px',
        'border': '1px solid black'
})

# generate popup text
info_box_template = """
<dl>
<dt>Station: {FIRE_STATION}</dt>
</dl>
<dl>
<dt>Count: {TOTAL COUNT}</dt>
</dl>
"""
count_info = [info_box_template.format(**row) for index, row in df_CAD_CALL_TYPE_STATION_GROUPBY_COPY.iterrows()]



# create marker layer
marker_station = gmaps.symbol_layer(locations_station,
                                    fill_color="blue",
                                    stroke_color="blue",
                                    scale=2,
                                    info_box_content=count_info)

# add heat_layer and symbols to figure
fig_station.add_layer(heat_layer)
fig_station.add_layer(marker_station)

# display the figure
display(fig_station)

# close the 2 jupyter gmap figures
fig.close()
fig_station.close()

## 5a. Interactive Google Maps Findings
- further diagrams could be done by analyzing sub-categories in FINAL_TYPE and CAD_CALL_TYPE but were not done due to cost concerns
- from a surface view it appears that the top 1000 geospatial locations that are serviced by TFS Fire Stations are right around them
    - no clear bias exists that indicates a missappropriation of TFS resources.

# 6. Jupyter Notebook References

[1] "Python Documentation."  *Python Software Foundation*.  [Online](https://docs.python.org/).  [Accessed August 04, 2020]

[2] G. Niemeyer.  "dateutil - powerful extensions to datetime."  *dateutil*.  [Online](https://github.com/dateutil/dateutil).  [Accessed August 04, 2020]

[3] "pandas."  *PyData*.  [Online](https://pandas.pydata.org/).  [Accessed August 04, 2020]

[4] "NumPy - The fundamental package for scientific computing with Python."  *NumPy*.  [Online](https://numpy.org/).  [Accessed August 04, 2020]

[5] "Matplotlib:  Visualization with Python."  *The Matplotlib Development team*.  [Online](https://matplotlib.org/).  [Accessed August 04, 2020]

[6] P. Bugnion.  "jupyter-gmaps:  Interactive Google maps in the IPython Notebook."  *jupyter-gmaps*.  [Online](https://github.com/pbugnion/gmaps).  [Accessed August 04, 2020]
